refactor: package depths
