+++ pytorch_optimizer/optimizer/adabound.py
@@ -1,22 +1,21 @@
 import math
+from typing import List
 
 import torch
 from torch.optim.optimizer import Optimizer
 
-from pytorch_optimizer.base_optimizer import BaseOptimizer
-from pytorch_optimizer.gc import centralize_gradient
-from pytorch_optimizer.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
-from pytorch_optimizer.utils import projection
+from pytorch_optimizer.base.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.base_optimizer import BaseOptimizer
 
 
-class AdamP(Optimizer, BaseOptimizer):
+class AdaBound(Optimizer, BaseOptimizer):
     """
-    Reference : https://github.com/clovaai/AdamP
+    Reference : https://github.com/Luolc/AdaBound
     Example :
-        from pytorch_optimizer import AdamP
+        from pytorch_optimizer import AdaBound
         ...
         model = YourModel()
-        optimizer = AdamP(model.parameters())
+        optimizer = AdaBound(model.parameters())
         ...
         for input, output in data:
           optimizer.zero_grad()
@@ -29,54 +28,56 @@ class AdamP(Optimizer, BaseOptimizer):
         self,
         params: PARAMETERS,
         lr: float = 1e-3,
+        final_lr: float = 1e-1,
         betas: BETAS = (0.9, 0.999),
+        gamma: float = 1e-3,
         weight_decay: float = 0.0,
-        delta: float = 0.1,
-        wd_ratio: float = 0.1,
-        use_gc: bool = False,
-        nesterov: bool = False,
+        weight_decouple: bool = True,
+        fixed_decay: bool = False,
+        amsbound: bool = False,
         adamd_debias_term: bool = False,
         eps: float = 1e-8,
     ):
-        """AdamP optimizer
+        """AdaBound
         :param params: PARAMETERS. iterable of parameters to optimize or dicts defining parameter groups
         :param lr: float. learning rate
+        :param final_lr: float. final learning rate
         :param betas: BETAS. coefficients used for computing running averages of gradient and the squared hessian trace
+        :param gamma: float. convergence speed of the bound functions
         :param weight_decay: float. weight decay (L2 penalty)
-        :param delta: float. threshold that determines whether a set of parameters is scale invariant or not
-        :param wd_ratio: float. relative weight decay applied on scale-invariant parameters compared to that applied
-            on scale-variant parameters
-        :param use_gc: bool. use gradient centralization
-        :param nesterov: bool. enables Nesterov momentum
+        :param weight_decouple: bool. the optimizer uses decoupled weight decay as in AdamW
+        :param fixed_decay: bool. fix weight decay
+        :param amsbound: bool. whether to use the AMSBound variant
         :param adamd_debias_term: bool. Only correct the denominator to avoid inflating step sizes early in training
         :param eps: float. term added to the denominator to improve numerical stability
         """
         self.lr = lr
         self.betas = betas
         self.weight_decay = weight_decay
+        self.weight_decouple = weight_decouple
+        self.fixed_decay = fixed_decay
         self.eps = eps
-        self.wd_ratio = wd_ratio
-        self.use_gc = use_gc
 
         self.validate_parameters()
 
         defaults: DEFAULTS = dict(
             lr=lr,
             betas=betas,
+            final_lr=final_lr,
+            gamma=gamma,
             weight_decay=weight_decay,
-            delta=delta,
-            wd_ratio=wd_ratio,
-            nesterov=nesterov,
+            amsbound=amsbound,
             adamd_debias_term=adamd_debias_term,
             eps=eps,
         )
         super().__init__(params, defaults)
 
+        self.base_lrs: List[float] = [group['lr'] for group in self.param_groups]
+
     def validate_parameters(self):
         self.validate_learning_rate(self.lr)
         self.validate_betas(self.betas)
         self.validate_weight_decay(self.weight_decay)
-        self.validate_weight_decay_ratio(self.wd_ratio)
         self.validate_epsilon(self.eps)
 
     @torch.no_grad()
@@ -88,6 +89,8 @@ class AdamP(Optimizer, BaseOptimizer):
                 state['step'] = 0
                 state['exp_avg'] = torch.zeros_like(p)
                 state['exp_avg_sq'] = torch.zeros_like(p)
+                if group['amsbound']:
+                    state['max_exp_avg_sq'] = torch.zeros_like(p)
 
     @torch.no_grad()
     def step(self, closure: CLOSURE = None) -> LOSS:
@@ -96,60 +99,60 @@ class AdamP(Optimizer, BaseOptimizer):
             with torch.enable_grad():
                 loss = closure()
 
-        for group in self.param_groups:
+        for group, base_lr in zip(self.param_groups, self.base_lrs):
             for p in group['params']:
                 if p.grad is None:
                     continue
 
                 grad = p.grad
                 if grad.is_sparse:
-                    raise RuntimeError('AdamP does not support sparse gradients')
+                    raise RuntimeError('AdaBound does not support sparse gradients')
 
                 state = self.state[p]
+
                 if len(state) == 0:
                     state['step'] = 0
                     state['exp_avg'] = torch.zeros_like(p)
                     state['exp_avg_sq'] = torch.zeros_like(p)
-
-                exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
+                    if group['amsbound']:
+                        state['max_exp_avg_sq'] = torch.zeros_like(p)
 
                 state['step'] += 1
+                exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
+
+                if self.weight_decouple:
+                    if not self.fixed_decay:
+                        p.mul_(1.0 - group['lr'] * group['weight_decay'])
+                    else:
+                        p.mul_(1.0 - group['weight_decay'])
+                else:
+                    if group['weight_decay'] != 0:
+                        grad.add_(p, alpha=group['weight_decay'])
+
                 beta1, beta2 = group['betas']
-
-                bias_correction1 = 1.0 - beta1 ** state['step']
-                bias_correction2 = 1.0 - beta2 ** state['step']
-
-                if self.use_gc:
-                    grad = centralize_gradient(grad, gc_conv_only=False)
 
                 exp_avg.mul_(beta1).add_(grad, alpha=1.0 - beta1)
                 exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1.0 - beta2)
 
-                de_nom = (exp_avg_sq.sqrt() / math.sqrt(bias_correction2)).add_(group['eps'])
+                if group['amsbound']:
+                    exp_avg_sq = torch.max(state['max_exp_avg_sq'], exp_avg_sq)
 
-                if group['nesterov']:
-                    perturb = (beta1 * exp_avg + (1.0 - beta1) * grad) / de_nom
-                else:
-                    perturb = exp_avg / de_nom
+                de_nom = exp_avg_sq.sqrt().add_(group['eps'])
 
-                wd_ratio: float = 1
-                if len(p.shape) > 1:
-                    perturb, wd_ratio = projection(
-                        p,
-                        grad,
-                        perturb,
-                        group['delta'],
-                        group['wd_ratio'],
-                        group['eps'],
-                    )
+                bias_correction1 = 1.0 - beta1 ** state['step']
+                bias_correction2 = 1.0 - beta2 ** state['step']
 
-                if group['weight_decay'] > 0:
-                    p.mul_(1.0 - group['lr'] * group['weight_decay'] * wd_ratio)
-
-                step_size = group['lr']
+                step_size = group['lr'] * math.sqrt(bias_correction2)
                 if not group['adamd_debias_term']:
                     step_size /= bias_correction1
 
-                p.add_(perturb, alpha=-step_size)
+                final_lr = group['final_lr'] * group['lr'] / base_lr
+                lower_bound = final_lr * (1 - 1 / (group['gamma'] * state['step'] + 1))
+                upper_bound = final_lr * (1 + 1 / (group['gamma'] * state['step']))
+
+                step_size = torch.full_like(de_nom, step_size)
+                step_size.div_(de_nom).clamp_(lower_bound, upper_bound).mul_(exp_avg)
+
+                p.add_(-step_size)
 
         return loss
