+++ pytorch_optimizer/optimizer/diffgrad.py
@@ -1,21 +1,20 @@
 import math
-from typing import List
 
 import torch
 from torch.optim.optimizer import Optimizer
 
-from pytorch_optimizer.base_optimizer import BaseOptimizer
-from pytorch_optimizer.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.base_optimizer import BaseOptimizer
 
 
-class AdaBound(Optimizer, BaseOptimizer):
+class DiffGrad(Optimizer, BaseOptimizer):
     """
-    Reference : https://github.com/Luolc/AdaBound
+    Reference : https://github.com/shivram1987/diffGrad
     Example :
-        from pytorch_optimizer import AdaBound
+        from pytorch_optimizer import DiffGrad
         ...
         model = YourModel()
-        optimizer = AdaBound(model.parameters())
+        optimizer = DiffGrad(model.parameters())
         ...
         for input, output in data:
           optimizer.zero_grad()
@@ -28,51 +27,30 @@ class AdaBound(Optimizer, BaseOptimizer):
         self,
         params: PARAMETERS,
         lr: float = 1e-3,
-        final_lr: float = 1e-1,
         betas: BETAS = (0.9, 0.999),
-        gamma: float = 1e-3,
-        weight_decay: float = 0.0,
-        weight_decouple: bool = True,
-        fixed_decay: bool = False,
-        amsbound: bool = False,
-        adamd_debias_term: bool = False,
         eps: float = 1e-8,
+        weight_decay: float = 0.0,
+        adamd_debias_term: bool = False,
     ):
-        """AdaBound
+        """DiffGrad optimizer
         :param params: PARAMETERS. iterable of parameters to optimize or dicts defining parameter groups
         :param lr: float. learning rate
-        :param final_lr: float. final learning rate
         :param betas: BETAS. coefficients used for computing running averages of gradient and the squared hessian trace
-        :param gamma: float. convergence speed of the bound functions
-        :param weight_decay: float. weight decay (L2 penalty)
-        :param weight_decouple: bool. the optimizer uses decoupled weight decay as in AdamW
-        :param fixed_decay: bool. fix weight decay
-        :param amsbound: bool. whether to use the AMSBound variant
-        :param adamd_debias_term: bool. Only correct the denominator to avoid inflating step sizes early in training
         :param eps: float. term added to the denominator to improve numerical stability
+        :param weight_decay: float. weight decay (L2 penalty)
+        :param adamd_debias_term: bool. Only correct the denominator to avoid inflating step sizes early in training
         """
         self.lr = lr
+        self.eps = eps
         self.betas = betas
         self.weight_decay = weight_decay
-        self.weight_decouple = weight_decouple
-        self.fixed_decay = fixed_decay
-        self.eps = eps
 
         self.validate_parameters()
 
         defaults: DEFAULTS = dict(
-            lr=lr,
-            betas=betas,
-            final_lr=final_lr,
-            gamma=gamma,
-            weight_decay=weight_decay,
-            amsbound=amsbound,
-            adamd_debias_term=adamd_debias_term,
-            eps=eps,
+            lr=lr, betas=betas, eps=eps, weight_decay=weight_decay, adamd_debias_term=adamd_debias_term
         )
         super().__init__(params, defaults)
-
-        self.base_lrs: List[float] = [group['lr'] for group in self.param_groups]
 
     def validate_parameters(self):
         self.validate_learning_rate(self.lr)
@@ -89,8 +67,7 @@ class AdaBound(Optimizer, BaseOptimizer):
                 state['step'] = 0
                 state['exp_avg'] = torch.zeros_like(p)
                 state['exp_avg_sq'] = torch.zeros_like(p)
-                if group['amsbound']:
-                    state['max_exp_avg_sq'] = torch.zeros_like(p)
+                state['previous_grad'] = torch.zeros_like(p)
 
     @torch.no_grad()
     def step(self, closure: CLOSURE = None) -> LOSS:
@@ -99,60 +76,50 @@ class AdaBound(Optimizer, BaseOptimizer):
             with torch.enable_grad():
                 loss = closure()
 
-        for group, base_lr in zip(self.param_groups, self.base_lrs):
+        for group in self.param_groups:
             for p in group['params']:
                 if p.grad is None:
                     continue
 
                 grad = p.grad
                 if grad.is_sparse:
-                    raise RuntimeError('AdaBound does not support sparse gradients')
+                    raise RuntimeError('diffGrad does not support sparse gradients')
 
                 state = self.state[p]
-
                 if len(state) == 0:
                     state['step'] = 0
                     state['exp_avg'] = torch.zeros_like(p)
                     state['exp_avg_sq'] = torch.zeros_like(p)
-                    if group['amsbound']:
-                        state['max_exp_avg_sq'] = torch.zeros_like(p)
+                    state['previous_grad'] = torch.zeros_like(p)
+
+                exp_avg, exp_avg_sq, previous_grad = state['exp_avg'], state['exp_avg_sq'], state['previous_grad']
+
+                if group['weight_decay'] != 0:
+                    grad.add_(p, alpha=group['weight_decay'])
 
                 state['step'] += 1
-                exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
-
-                if self.weight_decouple:
-                    if not self.fixed_decay:
-                        p.mul_(1.0 - group['lr'] * group['weight_decay'])
-                    else:
-                        p.mul_(1.0 - group['weight_decay'])
-                else:
-                    if group['weight_decay'] != 0:
-                        grad.add_(p, alpha=group['weight_decay'])
-
                 beta1, beta2 = group['betas']
 
+                # Decay the first and second moment running average coefficient
                 exp_avg.mul_(beta1).add_(grad, alpha=1.0 - beta1)
                 exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1.0 - beta2)
-
-                if group['amsbound']:
-                    exp_avg_sq = torch.max(state['max_exp_avg_sq'], exp_avg_sq)
-
                 de_nom = exp_avg_sq.sqrt().add_(group['eps'])
 
                 bias_correction1 = 1.0 - beta1 ** state['step']
                 bias_correction2 = 1.0 - beta2 ** state['step']
 
+                # compute diffGrad coefficient (dfc)
+                diff = abs(previous_grad - grad)
+                dfc = 1.0 / (1.0 + torch.exp(-diff))
+                state['previous_grad'] = grad.clone()
+
+                # update momentum with dfc
+                exp_avg1 = exp_avg * dfc
+
                 step_size = group['lr'] * math.sqrt(bias_correction2)
                 if not group['adamd_debias_term']:
                     step_size /= bias_correction1
 
-                final_lr = group['final_lr'] * group['lr'] / base_lr
-                lower_bound = final_lr * (1 - 1 / (group['gamma'] * state['step'] + 1))
-                upper_bound = final_lr * (1 + 1 / (group['gamma'] * state['step']))
-
-                step_size = torch.full_like(de_nom, step_size)
-                step_size.div_(de_nom).clamp_(lower_bound, upper_bound).mul_(exp_avg)
-
-                p.add_(-step_size)
+                p.addcdiv_(exp_avg1, de_nom, value=-step_size)
 
         return loss
