+++ pytorch_optimizer/optimizer/diffrgrad.py
@@ -3,19 +3,20 @@ import math
 import torch
 from torch.optim.optimizer import Optimizer
 
-from pytorch_optimizer.base_optimizer import BaseOptimizer
-from pytorch_optimizer.gc import centralize_gradient
-from pytorch_optimizer.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.types import BETAS, CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.base_optimizer import BaseOptimizer
 
 
-class Ranger(Optimizer, BaseOptimizer):
+class DiffRGrad(Optimizer, BaseOptimizer):
     """
-    Reference : https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer
+    Reference 1 : https://github.com/shivram1987/diffGrad
+    Reference 2 : https://github.com/LiyuanLucasLiu/RAdam
+    Reference 3 : https://github.com/lessw2020/Best-Deep-Learning-Optimizers/blob/master/diffgrad/diff_rgrad.py
     Example :
-        from pytorch_optimizer import Ranger
+        from pytorch_optimizer import DiffRGrad
         ...
         model = YourModel()
-        optimizer = Ranger(model.parameters())
+        optimizer = DiffRGrad(model.parameters())
         ...
         for input, output in data:
           optimizer.zero_grad()
@@ -28,47 +29,35 @@ class Ranger(Optimizer, BaseOptimizer):
         self,
         params: PARAMETERS,
         lr: float = 1e-3,
-        alpha: float = 0.5,
-        k: int = 6,
-        n_sma_threshold: int = 5,
-        betas: BETAS = (0.95, 0.999),
-        eps: float = 1e-5,
+        betas: BETAS = (0.9, 0.999),
         weight_decay: float = 0.0,
-        use_gc: bool = True,
-        gc_conv_only: bool = False,
+        n_sma_threshold: int = 5,
+        degenerated_to_sgd: bool = True,
         adamd_debias_term: bool = False,
+        eps: float = 1e-8,
     ):
-        """Ranger optimizer
+        """Blend RAdam with DiffGrad
         :param params: PARAMETERS. iterable of parameters to optimize or dicts defining parameter groups
         :param lr: float. learning rate
         :param betas: BETAS. coefficients used for computing running averages of gradient and the squared hessian trace
         :param weight_decay: float. weight decay (L2 penalty)
         :param n_sma_threshold: int. (recommended is 5)
-        :param use_gc: bool. use Gradient Centralization (both convolution & fc layers)
-        :param gc_conv_only: bool. use Gradient Centralization (only convolution layer)
+        :param degenerated_to_sgd: bool. degenerated to SGD
         :param adamd_debias_term: bool. Only correct the denominator to avoid inflating step sizes early in training
         :param eps: float. term added to the denominator to improve numerical stability
         """
         self.lr = lr
-        self.alpha = alpha
-        self.k = k
-        self.n_sma_threshold = n_sma_threshold
         self.betas = betas
         self.weight_decay = weight_decay
-        self.use_gc = use_gc
+        self.n_sma_threshold = n_sma_threshold
+        self.degenerated_to_sgd = degenerated_to_sgd
         self.eps = eps
-
-        self.gc_gradient_threshold: int = 3 if gc_conv_only else 1
 
         self.validate_parameters()
 
         defaults: DEFAULTS = dict(
             lr=lr,
-            alpha=alpha,
-            k=k,
-            step_counter=0,
             betas=betas,
-            n_sma_threshold=n_sma_threshold,
             eps=eps,
             weight_decay=weight_decay,
             adamd_debias_term=adamd_debias_term,
@@ -80,7 +69,6 @@ class Ranger(Optimizer, BaseOptimizer):
         self.validate_learning_rate(self.lr)
         self.validate_betas(self.betas)
         self.validate_weight_decay(self.weight_decay)
-        self.validate_lookahead_k(self.k)
         self.validate_epsilon(self.eps)
 
     @torch.no_grad()
@@ -92,8 +80,7 @@ class Ranger(Optimizer, BaseOptimizer):
                 state['step'] = 0
                 state['exp_avg'] = torch.zeros_like(p)
                 state['exp_avg_sq'] = torch.zeros_like(p)
-                state['slow_buffer'] = torch.empty_like(p)
-                state['slow_buffer'].copy_(p)
+                state['previous_grad'] = torch.zeros_like(p)
 
     @torch.no_grad()
     def step(self, closure: CLOSURE = None) -> LOSS:
@@ -109,13 +96,13 @@ class Ranger(Optimizer, BaseOptimizer):
 
                 grad = p.grad
                 if grad.is_sparse:
-                    raise RuntimeError('Ranger does not support sparse gradients')
+                    raise RuntimeError('AdaBelief does not support sparse gradients')
 
                 if grad.dtype in (torch.float16, torch.bfloat16):
                     grad = grad.float()
 
                 p_fp32 = p
-                if p.dtype in {torch.float16, torch.bfloat16}:
+                if p.dtype in (torch.float16, torch.bfloat16):
                     p_fp32 = p_fp32.float()
 
                 state = self.state[p]
@@ -123,24 +110,26 @@ class Ranger(Optimizer, BaseOptimizer):
                     state['step'] = 0
                     state['exp_avg'] = torch.zeros_like(p_fp32)
                     state['exp_avg_sq'] = torch.zeros_like(p_fp32)
-                    state['slow_buffer'] = torch.empty_like(p_fp32)
-                    state['slow_buffer'].copy_(p_fp32)
+                    state['previous_grad'] = torch.zeros_like(p_fp32)
                 else:
                     state['exp_avg'] = state['exp_avg'].type_as(p_fp32)
                     state['exp_avg_sq'] = state['exp_avg_sq'].type_as(p_fp32)
+                    state['previous_grad'] = state['previous_grad'].type_as(p_fp32)
 
-                exp_avg, exp_avg_sq = state['exp_avg'], state['exp_avg_sq']
-                beta1, beta2 = group['betas']
-
-                if self.use_gc and grad.dim() > self.gc_gradient_threshold:
-                    grad = centralize_gradient(grad, gc_conv_only=False)
+                exp_avg, exp_avg_sq, previous_grad = state['exp_avg'], state['exp_avg_sq'], state['previous_grad']
 
                 state['step'] += 1
+                beta1, beta2 = group['betas']
+
+                bias_correction1 = 1.0 - beta1 ** state['step']
 
                 exp_avg.mul_(beta1).add_(grad, alpha=1.0 - beta1)
                 exp_avg_sq.mul_(beta2).addcmul_(grad, grad, value=1.0 - beta2)
 
-                bias_correction1 = 1.0 - beta1 ** state['step']
+                # compute diffGrad coefficient (dfc)
+                diff = abs(previous_grad - grad)
+                dfc = 1.0 / (1.0 + torch.exp(-diff))
+                state['previous_grad'] = grad.clone()
 
                 buffered = group['buffer'][state['step'] % 10]
                 if state['step'] == buffered[0]:
@@ -148,10 +137,11 @@ class Ranger(Optimizer, BaseOptimizer):
                 else:
                     buffered[0] = state['step']
                     beta2_t = beta2 ** state['step']
-                    n_sma_max = 2 / (1 - beta2) - 1
-                    n_sma = n_sma_max - 2 * state['step'] * beta2_t / (1 - beta2_t)
+                    n_sma_max = 2.0 / (1.0 - beta2) - 1.0
+                    n_sma = n_sma_max - 2.0 * state['step'] * beta2_t / (1.0 - beta2_t)
                     buffered[1] = n_sma
-                    if n_sma > self.n_sma_threshold:
+
+                    if n_sma >= self.n_sma_threshold:
                         rt = math.sqrt(
                             (1 - beta2_t)
                             * (n_sma - 4)
@@ -165,25 +155,28 @@ class Ranger(Optimizer, BaseOptimizer):
                         step_size = rt
                         if not group['adamd_debias_term']:
                             step_size /= bias_correction1
-                    else:
+                    elif self.degenerated_to_sgd:
                         step_size = 1.0 / bias_correction1
+                    else:
+                        step_size = -1
 
                     buffered[2] = step_size
 
-                if group['weight_decay'] != 0:
-                    p_fp32.add_(p_fp32, alpha=-group['weight_decay'] * group['lr'])
+                if n_sma >= self.n_sma_threshold:
+                    if group['weight_decay'] != 0:
+                        p_fp32.add_(p_fp32, alpha=-group['weight_decay'] * group['lr'])
 
-                if n_sma > self.n_sma_threshold:
                     de_nom = exp_avg_sq.sqrt().add_(group['eps'])
-                    p_fp32.addcdiv_(exp_avg, de_nom, value=-step_size * group['lr'])
-                else:
+
+                    # update momentum with dfc
+                    p_fp32.addcdiv_(exp_avg * dfc.float(), de_nom, value=-step_size * group['lr'])
+                elif step_size > 0:
+                    if group['weight_decay'] != 0:
+                        p_fp32.add_(p_fp32, alpha=-group['weight_decay'] * group['lr'])
+
                     p_fp32.add_(exp_avg, alpha=-step_size * group['lr'])
 
-                p.copy_(p_fp32)
-
-                if state['step'] % group['k'] == 0:
-                    slow_p = state['slow_buffer']
-                    slow_p.add_(p - slow_p, alpha=self.alpha)
-                    p.copy_(slow_p)
+                if p.dtype in (torch.float16, torch.bfloat16):
+                    p.copy_(p_fp32)
 
         return loss
