+++ pytorch_optimizer/optimizer/madgrad.py
@@ -8,8 +8,8 @@ import math
 import torch
 from torch.optim import Optimizer
 
-from pytorch_optimizer.base_optimizer import BaseOptimizer
-from pytorch_optimizer.types import CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.types import CLOSURE, DEFAULTS, LOSS, PARAMETERS
+from pytorch_optimizer.base.base_optimizer import BaseOptimizer
 
 
 class MADGRAD(Optimizer, BaseOptimizer):
