+++ pytorch_optimizer/optimizer/sam.py
@@ -3,8 +3,8 @@ from typing import Dict
 import torch
 from torch.optim.optimizer import Optimizer
 
-from pytorch_optimizer.base_optimizer import BaseOptimizer
-from pytorch_optimizer.types import CLOSURE, DEFAULTS, PARAMETERS
+from pytorch_optimizer.base.types import CLOSURE, DEFAULTS, PARAMETERS
+from pytorch_optimizer.base.base_optimizer import BaseOptimizer
 
 
 class SAM(Optimizer, BaseOptimizer):
